Agent: move mimikatz collector to credential collectors
