+++ monkey/infection_monkey/credential_collectors/mimikatz_collector/pypykatz_handler.py
@@ -3,9 +3,7 @@ from typing import Any, Dict, List, NewType
 
 from pypykatz.pypykatz import pypykatz
 
-from infection_monkey.system_info.windows_cred_collector.windows_credentials import (
-    WindowsCredentials,
-)
+from .windows_credentials import WindowsCredentials
 
 CREDENTIAL_TYPES = [
     "msv_creds",
