+++ monkey/infection_monkey/credential_collectors/mimikatz_collector/windows_credentials.py
